Temp

+++ WebContent/interestsDeleteAction.jsp
@@ -21,26 +21,50 @@ String symbol = request.getParameter("ticker");
 %>
 
 <%
+	int delete_check = 0;
 	Connection conn = null;
 	PreparedStatement pstmt= null;
 	String sql = null;
 	ResultSet rs = null;
+	Class.forName("com.mysql.cj.jdbc.Driver");
+	String url = "jdbc:mysql://localhost:3306/STOCKIT?serverTimezone=UTC";
+	conn = DriverManager.getConnection(url,"root","0000");
 	try{
-		Class.forName("com.mysql.cj.jdbc.Driver");
-		String url = "jdbc:mysql://localhost:3306/STOCKIT?serverTimezone=UTC";
-		conn = DriverManager.getConnection(url,"root","0000");
-	
-		String SQL = "DELETE FROM interest where userId = ? AND symbol = ?";
-	
-		pstmt = conn.prepareStatement(SQL);
+
+		
+		String SQL2 = "SELECT count(*) FROM interest where userId = ? AND symbol = ?";	
+		pstmt = conn.prepareStatement(SQL2);
 		pstmt.setString(1, id);						
 		pstmt.setString(2, symbol);
-		pstmt.executeUpdate();
+		rs = pstmt.executeQuery();
+		
+		while(rs.next()){
+            delete_check = rs.getInt(1);		//중복체크
+			}
 		}catch(Exception e){
 			out.println("DB 연동 오류 입니다.:" +e.getMessage());	
 		}
 	
+	if(delete_check == 0){
+		script.println("<script>");
+		script.println("alert('관심종목으로 등록되있지 않은 종목입니다.')");		
+		script.println("history.back()");
+		script.println("</script>");
+		
+	}else{
+		delete_check = 0;
+		try{
+			String SQL = "DELETE FROM interest where userId = ? AND symbol = ?";
+		
+			pstmt = conn.prepareStatement(SQL);
+			pstmt.setString(1, id);						
+			pstmt.setString(2, symbol);
+			pstmt.executeUpdate();
+			}catch(Exception e){
+				out.println("DB 연동 오류 입니다.:" +e.getMessage());	
+			}
 	
+	}
 	script.println("<script>");
 	script.println("alert('관심종목에서 해제되었습니다.')");		
 	script.println("history.back()");
